% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admission.R
\name{admission}
\alias{admission}
\title{Admission Data}
\description{
Admission to a graduate school of 400 applications. Four variables are available
\describe{
 \item{admission}{binary variable indicating admission }
}
}
