% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.krr.R
\name{cv.krr}
\alias{cv.krr}
\title{Cross validation for Kernel Ridge Regresssion}
\usage{
cv.krr(y, X, k.folds = 10, lambda = 10^seq(-8, 8, length.out = 100), 
center = TRUE, scale = TRUE, ...)
}
\arguments{
\item{y}{a numeric vector of responses.}

\item{X}{a matrix of covariates.}

\item{k.folds}{the number of folds in k-fold cross validation. Default is \code{k.folds = 10}.}

\item{lambda}{the vector with lambdas to be used. Default is \code{lambda = 10^seq(-8, 8, length.out = 100)}.}

\item{center}{logical to indicate to center the predictor variables in \code{X} to have column sums zero. 
Default is \code{TRUE}.}

\item{scale}{logical to indicate to scale the predictor variables in \code{X} to have column variance 1. 
Default is \code{TRUE}.}

\item{...}{other parameters that may be passed on to \link{krr}.}
}
\value{
Returns a list of the class \code{cv.krr} with the following fields.
\item{\code{yhatu}}{a matrix the predicted out of sample values obtained by k-fold cross validation. 
The number of rows is the same as the number of rows of \code{X}  
and the number of columns is the same as the length of the vector \code{lambda}.}
\item{\code{lambda}}{the vector \code{lambda}.}
\item{\code{k.folds}}{number of folds.}
\item{\code{rmse}}{vector of root mean squared error comparing \code{y} with k-fold cross 
validated predictions \code{yhat}.}
\item{\code{lambda.min}}{the \code{lambda} with the smalles \code{rmse}.}
\item{\code{edf}}{effective degrees of freedom using \code{krr} with \code{lambda.min} on 
all observations.}
\item{\code{call}}{string with the call to this function.}
}
\description{
Performs k-fold cross validation for kernel ridge regression.
}
\examples{
data(longley)
X   <- as.matrix(longley[, c(1, 2, 4:7)])
y   <- longley[, 3]     # Unemployed
res <- cv.krr(y, X ,kernel.type = "RBF", lambda = 10^seq(-5, 5, length.out = 100))
}
