% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krr.R
\name{krr}
\alias{krr}
\title{Kernel Ridge Regresssion}
\usage{
krr(y, X, lambda = 10^seq(-8, 8, length.out = 100), 
kernel.type = c("RBF","linear","nonhompolynom"), kernel.degree = 2, 
kernel.RBF.sigma = 1, center = TRUE, scale = TRUE)
}
\arguments{
\item{y}{a numeric vector of responses.}

\item{X}{a matrix of covariates.}

\item{lambda}{the vector with lambdas to be used. Default is \code{lambda = 10^seq(-8, 8, length.out = 100)}.}

\item{kernel.type}{the type of kernel to be used. \code{RBF} defines the radial basis function 
or gaussian kernel (default), \code{linear} specifies the linear kernel, \code{nonhompolynom} specifies 
nonhomogeneous polynomial kernel.}

\item{kernel.degree}{the degree of the \code{nonhompolynom} kernel.}

\item{kernel.RBF.sigma}{nonnegative hyperparamter for the \code{RBF} kernel.}

\item{center}{logical to indicate to center the predictor variables in \code{X} to have column sums zero. 
Default is \code{TRUE}.}

\item{scale}{logical to indicate to scale the predictor variables in \code{X} to have column variance 1. 
Default is \code{TRUE}.}

\item{yu}{a numeric vector for out of sample predictions}

\item{Xu}{a matrix of out-of-sample covariates.}
}
\value{
Returns a list of the class \code{krr} with the following fields.
\item{\code{yhat}}{a matrix with the predicted values (in sample). Has the same length as \code{y} 
and the number of columns is the same as the length of the vector \code{lambda}.}
\item{\code{lambda}}{the vector \code{lambda}.}
\item{\code{edf}}{vector of the effective degrees of freedom corresponding to the  \code{lambda}.}
\item{\code{K}}{the kernel matrix computed using \code{X}).}
\item{\code{ei}}{list with the eigen \code{vectors} and \code{values} of \code{K}.}
\item{\code{meanX}}{vector with means of \code{X} if \code{center == TRUE} and 
\code{NULL} otherwise.}
\item{\code{stdX}}{vector with standard deviations of \code{X} if \code{scale == TRUE} and 
\code{NULL} otherwise.}
\item{\code{kernel.type}}{string with the specified kernel type.}
\item{\code{kernel.degree}}{degree of the kernel.}
}
\description{
Performs kernel ridge regression.
}
\examples{
data(longley)
ind <- as.logical(c(T,T,T,T,T,F,T,T,F,T,T,T,F,T,T,T))
X  <- as.matrix(longley[ind, c(1, 2, 4:7)])
y  <- longley[ind, 3] # Unemployed
Xu <- as.matrix(longley[!ind, c(1, 2, 4:7)])
yu <- longley[!ind, 3] # Unemployed
res <- krr(y, X, kernel.type = "RBF", lambda = 10^seq(-5, 5, length.out = 100))
plot(res, type = "insample")
}
