% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permtestPCA.R
\name{permtestPCA}
\alias{permtestPCA}
\title{Permutation Test for PCA}
\usage{
permtestPCA(X, nTests = 100, alpha = 0.05, center.data = TRUE, scale.data = TRUE, ...)
}
\arguments{
\item{X}{Numeric n by no of variables matrix on which PCA should be done.}

\item{nTests}{Number of permutations to be done. Default is 100.}

\item{alpha}{Confidence level.}

\item{center.data}{Indicates whether the column means need to be removed. Default is TRUE.}

\item{scale.data}{Indicates whether the column needs to have sum of squares n. Default is TRUE.}
}
\value{
A matrix with the lower and upper alpha/2 quantiles for each of the eigenvalues.
}
\description{
Does the permutation test for principal components analysis. The output is a scree 
plot with alpha/2 and 1 - alpha/2 confidence intervals for each of the eigenvalues.
}
\examples{
permtestPCA(gsp_share)
}
