% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.krr.R
\name{plot.cv.krr}
\alias{plot.cv.krr}
\title{Fit plot for cross validated kernel ridge regression}
\usage{
plot.cv.krr(res, type = c("rmse", "profile", "edf"), ...)
}
\arguments{
\item{res}{An object of the class \code{cv.krr}.}

\item{type}{plots the root mean squared error for \code{"rmse"}, provides a profile plot 
for each of out-of-sample predicted y value for \code{"profile"}, and a plot of effective 
degrees of freedom against \code{lambda} for \code{"edf"}.}

\item{...}{additional arguments that are passed on to plot.}
}
\description{
Makes a k-fold cross validated fit plot for kernel ridge regression for different lambda.
}
\examples{
data(longley)
X   <- as.matrix(longley[, c(1, 2, 4:7)])
y   <- longley[, 3]     # Unemployed
res <- cv.krr(y, X ,kernel.type = "RBF", lambda = 10^seq(-5, 5, length.out = 100))
plot(res)                    # for default rmse against lambda
plot(res, type = "profile")  # for profile lines per predicted object against lambda
plot(res, type = "edf")      # for edf against lambda
}
