% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.krr.R
\name{plot.krr}
\alias{plot.krr}
\title{Fit plot for kernel ridge regression}
\usage{
plot.krr(res, type = c( "insample", "out-of-sample"))
}
\arguments{
\item{res}{An object of the class krr.}

\item{type}{Make a plot of the \code{insample} predictions (default) or 
\code{out-of-sample} predictions per lambda.}
}
\value{
Returns a list of two items
\item{result}{matrix of with values of lambda and the root mean squared error (rmse).}
\item{lambda.min}{the values of lambda that had minimal root mean squared error x(rmse).}
}
\description{
Makes a fit plot for kernel ridge regression for different lambda.
}
