% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.krr.R
\name{predict.krr}
\alias{predict.krr}
\title{Prediction method for Kernel Ridge Regression}
\usage{
krr (y, X, yu = NULL, Xu = NULL, lambda = 10^seq(-8, 8, length.out = 100), 
kernel.type = c("RBF","linear","nonhompolynom"), kernel.degree = 2, kernel.RBF.sigma = 1,
center = TRUE, scale = TRUE)
}
\arguments{
\item{res}{an \code{krr} object returned by \code{krr}.}

\item{Xu}{a matrix of out-of-sample covariates.}
}
\value{
Returns a matrix the predicted out of sample values. 
The number of rows is the same as the number of rows of \code{Xu}  
and the number of columns is the same as the length of the vector \code{lambda}.
}
\description{
Does out-of-sample prediction for kernel ridge regression.
}
\examples{
data(longley)
ind <- as.logical(c(T,T,T,T,T,F,T,T,F,T,T,T,F,T,T,T))
X  <- as.matrix(longley[ind, c(1, 2, 4:7)])
y  <- longley[ind, 3] # Unemployed
Xu <- as.matrix(longley[!ind, c(1, 2, 4:7)])
yu <- longley[!ind, 3] # Unemployed
res <- krr(y, X, kernel.type = "RBF", lambda = 10^seq(-5, 5, length.out = 100))
yhatu <- predict(res, Xu)
}
